projects/storyweaver/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
stripe==7.4.0
Pillow==10.1.0
markdown==3.5.1
pandocfilters==1.5.0
pypandoc==1.11
gunicorn==21.2.0
Werkzeug==3.0.0
SQLAlchemy==2.0.23
